requirements.txt
Django==2.2.9
django-cors-headers==2.5.0
djangorestframework==3.11.0
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
